Relocated the mail helper classes and updated namespace
