+++ composer.json
@@ -18,7 +18,7 @@
   },
   "type": "library",
   "autoload": {
-    "psr-4": {"Helpers\\": "lib/Helpers/"},
+    "psr-4": {"SendGrid\\": "lib/Helpers/"},
     "files": [
       "lib/SendGrid.php"
     ]
