+++ lib/Helpers/Mail/Footer.php
@@ -11,11 +11,13 @@
  * @link      http://packagist.org/packages/sendgrid/sendgrid
  */
 
-namespace Helpers;
+namespace SendGrid\Mail;
 
-class SandBoxMode implements \JsonSerializable
+class Footer implements \JsonSerializable
 {
     private $enable;
+    private $text;
+    private $html;
 
     public function setEnable($enable)
     {
@@ -26,11 +28,34 @@ class SandBoxMode implements \JsonSerializable
     {
         return $this->enable;
     }
+
+    public function setText($text)
+    {
+        $this->text = $text;
+    }
+
+    public function getText()
+    {
+        return $this->text;
+    }
+
+    public function setHtml($html)
+    {
+        $this->html = $html;
+    }
+
+    public function getHtml()
+    {
+        return $this->html;
+    }
+
     public function jsonSerialize()
     {
         return array_filter(
             [
-                'enable' => $this->getEnable()
+                'enable' => $this->getEnable(),
+                'text'   => $this->getText(),
+                'html'   => $this->getHtml()
             ],
             function ($value) {
                 return $value !== null;
