+++ lib/Helpers/Mail/SubscriptionTracking.php
@@ -11,7 +11,7 @@
  * @link      http://packagist.org/packages/sendgrid/sendgrid
  */
 
-namespace Helpers;
+namespace SendGrid\Mail;
 
 class SubscriptionTracking implements \JsonSerializable
 {
